+++ src/lib/lforms/lhc-common-utils.js
@@ -3,6 +3,7 @@
  */
 import moment from "moment"
 import copy from "fast-copy";
+import deepEqual from "deep-equal";
 
 const CommonUtils = {
 
@@ -296,12 +297,24 @@ const CommonUtils = {
   },
 
 
-// TODO: move to lhc-form-utils.js
+  /**
+   * deep copy of an object
+   * @param {*} sourceObj an object
+   * @returns an copied new object
+   */
   deepCopy: function(sourceObj) {
-    //console.dir(sourceObj);
-
-    //return jQuery.extend(true, {}, sourceObj);
     return copy(sourceObj);
+  },
+
+  
+  /**
+   * deep comparison of two objects
+   * @param {*} obj1 an object
+   * @param {*} obj2 an object
+   * @returns 
+   */
+  deepEqual: function(obj1, obj2) {
+    return deepEqual(obj1, obj2);
   },
 
 
@@ -311,12 +324,55 @@ const CommonUtils = {
    * @param msg the message to show
    * @param item (optional) the item in the form to which the message applies.
    */
-   showWarning: function(msg, item) {
+  showWarning: function(msg, item) {
     if (item)
       msg = 'The question "'+item.text+'" produced the following warning:  '+msg;
     console.log(msg);
     // TBD: add a warning visible on the page.
+  },
+  
+
+  /**
+   * Changes the answer's display text when there is a label and/or a score
+   * @param {*} answers the answers on an item
+   * @returns answers with a modified display text on each answer item, and
+   *          a flag whether to add sequence number for the each answer's displayed text.
+   */
+  getAnswerDisplayTextWithLabelAndScore: function(answers) {
+    // reset the modified answers (for the display text)
+    var modifiedAnswers = [];
+    var hasOneAnswerLabel = false;
+    var hasOneNumericAnswer = false;
+    if (answers && Array.isArray(answers)) {
+      for (var i = 0, iLen = answers.length; i < iLen; i++) {
+        var answerData = CommonUtils.deepCopy(answers[i]);
+  
+        var displayText = answerData.text;
+        // label is a string
+        if (answerData.label) {
+          displayText = answerData.label + ". " + displayText;
+          hasOneAnswerLabel = true;
+        }
+        // check if one of the values is numeric
+        else {
+          if (!hasOneNumericAnswer && !isNaN(answerData.text)) {
+            hasOneNumericAnswer = true;
+          }
+        }
+  
+        if (answerData.score !== undefined && answerData.score !== null)
+          displayText = displayText + " - " + answerData.score;
+        // always uses _displayText in autocomplete-lhc and radio buttons/checkboxes for display
+        answerData._displayText = displayText;
+        modifiedAnswers.push(answerData);
+      }  
+    }
+    // add seq num when there is no labels and no numeric values as answer
+    var acAddSeq = !hasOneAnswerLabel && !hasOneNumericAnswer;
+    
+    return [modifiedAnswers, acAddSeq];
   }
+
   
 };